% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_mpas.R
\docType{data}
\name{all_mpas}
\alias{all_mpas}
\title{Marine Protected Areas (MPAs) of Mexico}
\format{
A simple feature collection with 24 features and 5 fields
\describe{
\item{NOMBRE}{Name of the MPA in Spanish}
\item{CAT_DECRET}{Decree category, which define the type of MPA}
\item{ESTADOS}{State that have jurisdiction on the MPA}
\item{MUNICIPIOS}{Municipality that have jurisdiction on the MPA}
\item{REGION}{General regional localization of the MPA (in Spanish)}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{http://sig.conanp.gob.mx/website/pagsig/info_shape.htm}
}
\usage{
all_mpas
}
\description{
A \code{sf} object containing shapefiles of MPA polygons in Mexico
}
\keyword{datasets}
