% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings.R
\docType{data}
\name{landings}
\alias{landings}
\title{Catch data from the vessels in Mexico}
\format{
A \code{data.frame}  with 111,369 rows and 7 columns
\describe{
\item{office_name}{Name of office where the catch was reported to}
\item{year}{year when the catch was taken}
\item{category}{vessel category MENORES = artesanal vessels, MAYORES = industrial vessels}
\item{vessel_name}{Official name of the vessel}
\item{species_name}{General name of the species that was caught}
\item{live_catch_ton}{Fresh weight of the catch in tons}
\item{catch_ton}{final weight of the catch in tons}
...
}
}
\source{
\url{https://datos.gob.mx/busca/organization/conapesca}
}
\usage{
landings
}
\description{
A \code{data.frame} object containing catch data per each vessel from 2015 to 2020
}
\keyword{datasets}
