% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vms.R
\name{model_vms}
\alias{model_vms}
\title{Vessel Modeling with Gaussian Mixture Models}
\usage{
model_vms(df)
}
\arguments{
\item{df}{a data.frame preprocessed using the \code{preprocessing_vms()} function from this package}
}
\value{
a data.frame with a \code{vessel_state} column with the type of model implemented
}
\description{
This function uses \code{normalmixEM} from the \code{mixtools} package to model speed of vessels and estimates their behavior.
Specifically, if the vessel was in a fishing activity or cruising
}
\examples{
\dontrun{
model_vms(df)
}
}
