% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpas_buffers.R
\docType{data}
\name{mpas_buffers}
\alias{mpas_buffers}
\title{Buffer around remote Marine Protected Areas of Mexico}
\format{
A simple feature collection with 5 features and 2 fields
\describe{
\item{Name}{Name of the MPAs to which the buffer correspond}
\item{Description}{empty}
\item{geometry}{column containing geometry details}
...
}
}
\source{
this project
}
\usage{
mpas_buffers
}
\description{
A \code{sf} object containing shapefiles of buffers around remote MPAs in Mexico.
The buffer equals the area inside each MPA polygon and was created to assess differences in fishing
activity inside or outside each of the remote MPAs.
}
\keyword{datasets}
