% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_coastline.R
\docType{data}
\name{mx_coastline}
\alias{mx_coastline}
\title{Mexican coastline}
\format{
A simple feature collection with 177 features and 3 fields
\describe{
\item{featurecla}{Name of the object}
\item{scalerank}{resolution rank}
\item{min_zoom}{zoom precision}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://cran.r-project.org/web/packages/rnaturalearth/README.html}
}
\usage{
mx_coastline
}
\description{
A \code{sf} object containing a the Mexican coastline shapefile
}
\keyword{datasets}
