% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_coastline_buffer.R
\docType{data}
\name{mx_coastline_buffer}
\alias{mx_coastline_buffer}
\title{Buffer around the Mexican coastline}
\format{
A simple feature collection with 1 feature and 3 fields
\describe{
\item{featurecla}{Name of the object}
\item{scalerank}{resolution rank}
\item{min_zoom}{zoom precision}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://cran.r-project.org/package=rnaturalearth}
}
\usage{
mx_coastline_buffer
}
\description{
A \code{sf} object containing a buffer around Mexican coastline
that was used to create the inland shapefile available in this package.
}
\keyword{datasets}
