% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_eez.R
\docType{data}
\name{mx_eez}
\alias{mx_eez}
\title{Mexico shape}
\format{
A simple feature collection with 1 features and 2 fields
\describe{
\item{Name}{empty}
\item{Description}{empty}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://en.www.inegi.org.mx/}
}
\usage{
mx_eez
}
\description{
A \code{sf} object containing the shapefile representing Mexico
}
\keyword{datasets}
