% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_eez_pacific.R
\docType{data}
\name{mx_eez_pacific}
\alias{mx_eez_pacific}
\title{Economic Exclusive Zone (EEZ) of the Pacific side of Mexico}
\format{
A simple feature collection with 1 feature and 1 field
\describe{
\item{Name}{Mexican Pacific Exclusive Economic Zone}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://en.www.inegi.org.mx/}
}
\usage{
mx_eez_pacific
}
\description{
A \code{sf} object containing shapefiles of Mexican EEZ in the Pacific
}
\keyword{datasets}
