% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_inland.R
\docType{data}
\name{mx_inland}
\alias{mx_inland}
\title{Area inland of Mexico}
\format{
A simple feature collection with 1 feature and 2 fields
\describe{
\item{Name}{Mexico}
\item{Desciption}{empty}
\item{geometry}{column containing geometry details}
...
}
}
\source{
modified from Mexican shapefile
}
\usage{
mx_inland
}
\description{
A \code{sf} object containing shapefiles of inland area in Mexico
}
\keyword{datasets}
