% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_ports.R
\docType{data}
\name{mx_ports}
\alias{mx_ports}
\title{Ports and Marinas of Mexico}
\format{
A simple feature collection with 237 features and 2 fields
\describe{
\item{class}{Type of infrastructure it can be Puerto (Port), or Marina}
\item{name}{Name of the infrastructure (i.e. port or marina)}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://en.www.inegi.org.mx/}
}
\usage{
mx_ports
}
\description{
A \code{sf} object containing points representing the locations of Ports and Marinas in Mexico
}
\keyword{datasets}
