% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_shape.R
\docType{data}
\name{mx_shape}
\alias{mx_shape}
\title{Mexico mainland}
\format{
A simple feature collection with 1 feature and 2 fields
\describe{
\item{Name}{Mexico}
\item{Description}{empty}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{https://en.www.inegi.org.mx/}
}
\usage{
mx_shape
}
\description{
A \code{sf} object containing a shapefile of Mexico
}
\keyword{datasets}
