% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelagic_vessels_permits.R
\docType{data}
\name{pelagic_vessels_permits}
\alias{pelagic_vessels_permits}
\title{List of vessels with pelagic fishing permits}
\format{
A \code{data.frame} with 719 rows and 2 columns.
\describe{
\item{RNP}{Unique code identifing the vessel}
\item{vessel_name}{Name of the vessel}
...
}
}
\source{
\url{www.datamares.org}
}
\usage{
pelagic_vessels_permits
}
\description{
A \code{data.frame} object extracted from a raw dataset of permits available under request at dataMares (https://datamares.org/)
}
\keyword{datasets}
