% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_mpas.R
\docType{data}
\name{remote_mpas}
\alias{remote_mpas}
\title{Remote Marine Protected Areas (MPAs) of Mexico}
\format{
A simple feature collection with 5 features and 2 fields
\describe{
\item{Name}{Name of the remote MPA in Spanish}
\item{Description}{empty}
\item{geometry}{column containing geometry details}
...
}
}
\source{
\url{http://sig.conanp.gob.mx/}
}
\usage{
remote_mpas
}
\description{
A \code{sf} object containing shapefiles of remote MPA polygons in Mexico that are of particular
conservation interest
}
\keyword{datasets}
