% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_download.R
\name{vms_download}
\alias{vms_download}
\title{Download VMS raw data}
\usage{
vms_download(year = lubridate::year((Sys.time())) - 1)
}
\arguments{
\item{year}{year of data that user wants to download are selected default to the last year. A vector of years can also be used.}
}
\value{
saves downloaded data into a folder called \code{VMS-data} within the working directory
}
\description{
This functions download data form the datos abiertos inititative.
}
\details{
Data are downloaded from this link: https://www.datos.gob.mx/busca/dataset/localizacion-y-monitoreo-satelital-de-embarcaciones-pesqueras/resource/309e872a-dbca-4962-b14f-f0da833abebe
Downloaded data are stored automatically in a \code{VMS-data} folder. Within the main folder, data is organized in
different folders by months (in spanish names) and within each there are multiple \code{.csv} files
each containing two weeks of data points.
}
\examples{

# Download single year
\dontrun{
vms_download(2019)
}

# Download multiple years
\dontrun{
vms_download(2015:2021)
}

}
